refactor: abstract it into smaller and simpler modules
